pyproject.toml
[tool.poetry]
name = "checker"
version = "0.1.0"
description = "Serverless Endpoint checker"
authors = ["Momfer"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"


[tool.poetry.group.cdk.dependencies]
aws-cdk-lib = "==2.110.1"
pydantic = "^2.3.0"
pydantic-settings = "^2.0.3"
poetry-plugin-export = "^1.6.0"
aws-cdk-aws-lambda-python-alpha = "^2.110.1a0"


[tool.poetry.group.lambda.dependencies]
boto3 = "^1.28.49"
botocore = "^1.32.6"
mypy-boto3-dynamodb = "^1.29.0"
aws-lambda-typing = "^2.18.0"
pydantic = "^2.3.0"
pydantic-core = "^2.14.5"


[tool.poetry.group.tst.dependencies]
cdk-nag = "^2.27.130"
pytest = "^7.4.1"


[tool.poetry.group.dev.dependencies]
boto3-stubs = { extras = ["all"], version = "^1.29.0" }
ruff = "^0.0.287"
httpie = "^3.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pyright]
typeCheckingMode = "strict"

exclude = [
    "**/node_modules",
    "**/__pycache__",
    "venv/",
    ".venv/"
]


[tool.isort]
profile = "black"


[tool.black]
line-length = 100


[tool.ruff]
# E & F - defaults
# W     - optional warnings (for example trailing whitespace)
# C90   - code complexity
# ANN   - enforce strict type annotation
# G     - should not have f-strings in logs
select = ["E", "F", "W", "C90", "ANN", "G"]
logger-objects = ["logger.log"] # <- Change path to logger object
line-length = 100
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    ".venv",
]
